+++ src/controller/order.js
@@ -40,6 +40,17 @@ class OrderController {
             res.status(err.status? err.status: 500).send(errs)
         }
     }
+    async checkout(req, res) {
+        try {
+            const { medicineOrderIds } = req.body
+            const { userId } = req
+            const response = await this.orderService.checkout(userId, medicineOrderIds)
+            res.status(200).send(response)
+        } catch (err) {
+            const errs = [new ErrorResponse(ErrorType.ERROR_MEDICINE_ORDER_FETCH, err.message)]
+            res.status(err.status? err.status: 500).send(errs)
+        }
+    }
 }
 
 module.exports = { OrderController }
